+++ digits_3d/utils.py
@@ -1,7 +1,5 @@
 import os
-
 import numpy as np
-import pandas as pd
 from matplotlib import pyplot as plt
 import seaborn as sns
